+++ changelog.md
@@ -14,6 +14,7 @@
   - `languages.total.bytes` &rarr; `languages.bytes`
   - `languages.total.unknownBytes` &rarr; `unknown.bytes`
 - **Breaking:** Changed the default CLI output format from JSON to summary.
+- **Breaking:** Changed CLI array arguments to require delimiters to be spaces instead of other characters.
 - Added CLI option `--json` to display the results as JSON (the previous default behavior).
 - Removed CLI option `--summary` as it is now the default output format.
 - Removed CLI option `--files` to list all files parsed.